pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "langchain>=1.2.15",
    "langchain-anthropic>=1.4.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.12",
    "mcp>=1.27.0",
    "pydantic>=2.12.5",
    "pillow>=12.2.0",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "rich>=14.3.3",
    "tavily-python>=0.7.23",
]
